requirements.txt
aniso8601==9.0.1
biotite==0.39.0
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-RESTful==0.3.10
Flask-Session==0.5.0
gunicorn==21.2.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
msgpack==1.0.7
networkx==3.2.1
numpy==1.26.3
packaging==23.2
pandas==2.1.4
plotly==5.18.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
tenacity==8.2.3
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
wheel==0.41.2